refactor TTableEntry
The class used to be mutable in an uncontrolled way, now properties only
expose getters, while setting the values is only possible via the
provided update() function. This allows some safety against unintended
modification of an entry's state while allowing for the same
optimisation, i.e. reuse of the entries in a transposition table, where
multiple milions of entries will land, but only up to a fixed maximum
number of entry objects will actually be created.