include status of the game in engine response
Status can be ONGOING or FINISHED. Sealed class is used, so FINISHED can
have additional properties, predominantly an information of who won, AI
or HUMAN.

Also refactored queryProcessor.kt in order to keep the code readable
despite the addition of the new functionality. Reduced returns to just
one and added a getStatus() extension to GameState in order to capture
the state in an elegant way.