work on a copy when processing engine query
When user provides a query for the library to consume, the negamax
algorithm is now run with a copy of GameState that the user has. This
allows the user to treat the object as if it were immutable.

This fixes a range of possible bugs in apps using the library, most
notably the app could assume the state of the object doesn't change
after it's passed over to the engine and for example attempt saving it
on a different thread than the library is mutating the object. This
would result in valuable but unexpected insights into the way the engine
works through possible paths the game can take, but would also corrupt
game saves.