improve evaluation
Evaluation didn't see the end game opportunities as valuable, instead
the engine, when victory was pretty much assured, engaged in further
maximization of score, potentially leading to an endless play-with-food
condition.
Evaluation function was modified to address this issue. Now sums are
tracked separately for each player and then added together to obtain the
evaluation of a given GameState.
Also the longer (in moves count) it takes to win, the lower the
evaluation.

Updated tests to match the new intended behaviour.