code cleanup
replaced Int.MIN_VALUE and MAX_VALUE with MIN_SCORE and MAX_SCORE
constants, in order to improve readability. The Int properties had to be
used with immediate subtraction or addition of 1, because these values
will be reverted and the absolute values of upper and lower limit are
different, thus one can overflow when a minus sign is applied to it.
This is now handled at the time of constant declaration, once.

Removed unused import statement in IntExtensionsTest.