extract incremental evaluation from GameState
GameState is now provided a factory of incremental evaluators at
construction time. The factory is then used to get an instance of an
IncrementalEvaluator.

Also moved destination value calculation from move application over to
Move class, so it can be reused in the evaluation and using the
Evaluator is easier (you no longer need to call it before/after the move
application in GameState. Evaluator still holds instance of the
gameState it serves and usees it to obtain movesCount.