code cleanup
Moved entirely to the new, incremental evaluation. This is primarily
because having two implementations of the same feature, working side by
side, is problematic, as both implementations would have to be
maintained. Now the old implementation is completely gone and all parts
of code use the new one instead.

Moved responsibilities related to evaluation to GameState class, so when
evaluation is needed, it's obtained via a method on a GameState
instance.

Renamed getEvaluation() to evaluate() in IncrementalEvaluator and
in the concrete class implementing it.

Fixed wrong package set for BestMovePickerTest.

Fixed bugs in tests. The bugs manifested themselves due to a slight
difference in the way input is processed by the new evaluation
implementation. The old code didn't rely on the parts of the imput that
were corrupted in the tests, therefore the corruption went unnoticed.