introduce killer heuristic
Implemented a helper class KillerHeuristic as well as added tests and
ensured proper function of the new class. Modified move ordering to
allow for indicating known good non-capture moves, if any.
Added KillerHeuristic parameter in negamax function and in
the bestMovePicker.kt's functions. Also included absolute depth so the
KillerHeuristic can be wired in snugly.