expect query to contain state with move applied
GameState in query provided to the engine must now have the chosen move
applied on the library user's side. This simplifies storage options for
the users, as when the last move was by human player, the move chosen
does not have to be saved separately to be re-submitted to the engine.
The precise move chosen is inconsequential in and out of itself - only
the state is produces on the board matters, so applying it immediately
by the user and saving the resulting state is perfectly sufficient.