use more performant ways of obtaining current time
It was revealed by a profiler that on JVM the library spends a
significant time just obtaining the current time in order to know when
to stop iterative deepening search. The cause turned out to be use of
Calendar.getInstance().timeInMillis which is less performant than the
System.currentTimeMillis() used instead.
The fix was applied and large performance gain was
observed on JVM when time constraint is enabled.

For JS, Date().getTime() was replaced with Date.now().