reverse end checks order in negamax
In the beginning of the negamax(), two checks are performed. One to
establish whether or not depth left to search is 0, another to ensure
there still is time left till deadline.
Since time checks are more expensive than depthLeft == 0 checks, the
latter are now checked first. This reduces the number of deadline checks
in cases of low max depth allowed, either because of the search being
shallow as a whole or during iterative deepening, as the current
iteration's max depth is reached per each iteration. Since evaluation is
incremental, its efficient enough to be favoured over the time check
too.