simplify performance testing
Since queryProcessor no longer takes a Move parameter, instead expecting
it to be already applied to GameState passed as another parameter, it
made sense to remove the redundant explicit move information from the
test cases. The exact last move doesn't make a difference to the engine
anymore - it only receives the exact state passed to it.
Move counters were incremented by 1 to reflect the fact the moves were
already applied.