add humanPlayerScore property to GameState
This property provides users of the library with a way to access raw
materiel difference between players - a de facto result of the game. The
more radical the difference, the more crushing the victory of one of the
contestants. This is more useful for auto-balancing than the actual
evaluation internally used in the engine, as it doesn't radically jump
up to the higher end of the integer range.

The evaluation is no longer accessible (were marked internal).