implement basics of the parallel solution
The solution is non-working at this stage, because of lack of support
for runBlocking in the common module of multiplatform library projects.

This is because implementing runBlocking for JS platform would be
problematic, so coroutines don't support it currently.

Once the support is provided, or an elegant work-around is found, this
can be revisited and done in the same kind of way as in the
experiments/coroutinesTest.kt.